requirements.txt
anyascii==0.3.3
arrow==1.3.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.14.2
binaryornot==0.4.4
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.8
cookiecutter==2.6.0
defusedxml==0.7.1
dj-database-url==3.0.1
django==4.2.25
django-allauth==65.13.0
django-browser-reload==1.15.0
django-filter==24.3
django-otp==1.5.4
django-permissionedforms==0.1
django-ratelimit==4.1.0
django-stubs-ext==5.1.3
django-taggit==6.1.0
django-tailwind==3.6.0
django-tasks==0.6.1
djangorestframework==3.15.2
easypost==9.5.0
et-xmlfile==2.0.0
Faker==35.2.2
filetype==1.2.0
gunicorn==23.0.0
idna==3.10
jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openpyxl==3.1.5
packaging==25.0
phonenumbers==9.0.18
pillow==10.4.0
pillow-heif==0.18.0
psutil==7.1.3
psycopg2-binary==2.9.10
pygments==2.19.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
qrcode==7.4.2
requests==2.32.4
rich==14.1.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stripe==13.0.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.13.2
urllib3==2.2.3
whitenoise==6.7.0
APScheduler==3.10.4

# Code quality and linting
black==24.8.0
flake8==7.1.1
isort==5.13.2
mypy==1.14.1
pylint==3.2.7
pre-commit==3.5.0
django-stubs==5.1.2
djlint==1.35.2
